Route cart summary API, drop dead code
